Fix some of the bugs in the Python version of the sudoku vis code.
This adds features to the framework class to stop lots of failures in
GLUT callback handlers from overwhelming the console; only the first
time a particular exception happens will it be printed out.

This corrects the argument handling and fixes a number of other small
bugs in the visualiser code itself.